move NetworkBase into its own file
